discord-bot/requirements.txt
mysql-connector-python==8.3.0
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
APScheduler==3.10.4
